% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histredline.R
\name{histredline}
\alias{histredline}
\title{Histogram with red mean line}
\usage{
histredline(x, y)
}
\arguments{
\item{x}{Dataset which contains the numeric variable}

\item{y}{Variable used to generate the histogram}
}
\value{
A histogram with a red line denoting the mean
}
\description{
This function creates a histogram of a numeric variable and adds a red line denoting the position of the mean in the distribution of values of the variable.
}
